+++ packages/client/src/tx/types.ts
@@ -91,7 +91,8 @@ export type CustomSignedExtensionValues =
       additionalSigned: any
     }
 
-export type TxOptions<Asset> = Partial<
+//  = CustomSignedExtensionValues
+export type TxOptions<Asset, Ext> = Partial<
   void extends Asset
     ? {
         /**
@@ -122,7 +123,7 @@ export type TxOptions<Asset> = Partial<
          * its dynamic codecs. At least one of the 2 values must be included
          * into the signed-extension Object.
          */
-        customSignedExtensions: Record<string, CustomSignedExtensionValues>
+        customSignedExtensions: Ext
       }
     : {
         /**
@@ -153,7 +154,7 @@ export type TxOptions<Asset> = Partial<
          * its dynamic codecs. At least one of the 2 values must be included
          * into the signed-extension Object.
          */
-        customSignedExtensions: Record<string, CustomSignedExtensionValues>
+        customSignedExtensions: Ext
         /**
          * Asset information to pay fees, tip, etc. By default it'll use the
          * native token of the chain.
@@ -231,14 +232,14 @@ export type OfflineTxExtensions<Asset> = void extends Asset
       asset?: Asset
     }
 
-export type TxPromise<Asset> = (
+export type TxPromise<Asset, Ext> = (
   from: PolkadotSigner,
-  txOptions?: TxOptions<Asset>,
+  txOptions?: TxOptions<Asset, Ext>,
 ) => Promise<TxFinalizedPayload>
 
-export type TxObservable<Asset> = (
+export type TxObservable<Asset, Ext> = (
   from: PolkadotSigner,
-  txOptions?: TxOptions<Asset>,
+  txOptions?: TxOptions<Asset, Ext>,
 ) => Observable<TxEvent>
 
 export interface TxCall {
@@ -250,9 +251,19 @@ export interface TxCall {
   (): Promise<Binary>
 }
 
-export type TxSignFn<Asset> = (
+export interface TxBare {
+  /**
+   * SCALE-encoded Bare (aka Unsigned) transaction ready to be broadcasted.
+   *
+   * @returns Promise resolving in a Bare transaction (it falls back to an
+   *          unsigned transaction if only v4 is available)
+   */
+  (): Promise<HexString>
+}
+
+export type TxSignFn<Asset, Ext> = (
   from: PolkadotSigner,
-  txOptions?: TxOptions<Asset>,
+  txOptions?: TxOptions<Asset, Ext>,
 ) => Promise<HexString>
 
 export type PaymentInfo = {
@@ -273,6 +284,7 @@ export type InnerTransaction<
   Pallet extends string,
   Name extends string,
   Asset,
+  Ext,
 > = {
   /**
    * Pack the transaction, sends it to the signer, and return the signature
@@ -283,7 +295,7 @@ export type InnerTransaction<
    * @param txOptions  Optionally pass any number of txOptions.
    * @returns Encoded `SignedExtrinsic` ready for broadcasting.
    */
-  sign: TxSignFn<Asset>
+  sign: TxSignFn<Asset, Ext>
   /**
    * Observable-based all-in-one transaction submitting. It will sign,
    * broadcast, and track the transaction. The observable is singlecast, i.e.
@@ -294,7 +306,7 @@ export type InnerTransaction<
    * @param txOptions  Optionally pass any number of txOptions.
    * @returns Observable to the transaction.
    */
-  signSubmitAndWatch: TxObservable<Asset>
+  signSubmitAndWatch: TxObservable<Asset, Ext>
   /**
    * Pack the transaction, sends it to the signer, broadcast, and track the
    * transaction. The promise will resolve as soon as the transaction in found
@@ -305,11 +317,16 @@ export type InnerTransaction<
    * @param txOptions  Optionally pass any number of txOptions.
    * @returns Finalized transaction information.
    */
-  signAndSubmit: TxPromise<Asset>
+  signAndSubmit: TxPromise<Asset, Ext>
   /**
    * SCALE-encoded callData of the transaction.
    */
   getEncodedData: TxCall
+  /**
+   * SCALE-encoded Bare (aka Unsigned) transaction ready to be broadcasted.
+   */
+  getBareTx: TxBare
+
   /**
    * Estimate fees against the latest known `finalizedBlock`
    *
@@ -319,7 +336,7 @@ export type InnerTransaction<
    */
   getEstimatedFees: (
     from: Uint8Array | SS58String,
-    txOptions?: TxOptions<Asset>,
+    txOptions?: TxOptions<Asset, Ext>,
   ) => Promise<bigint>
   /**
    * Payment info against the latest known `finalizedBlock`
@@ -331,7 +348,7 @@ export type InnerTransaction<
    */
   getPaymentInfo: (
     from: Uint8Array | SS58String,
-    txOptions?: TxOptions<Asset>,
+    txOptions?: TxOptions<Asset, Ext>,
   ) => Promise<PaymentInfo>
 
   /**
@@ -346,12 +363,19 @@ export type Transaction<
   Pallet extends string,
   Name extends string,
   Asset,
-> = InnerTransaction<Arg, Pallet, Name, Asset>
+  Ext = Record<string, CustomSignedExtensionValues>,
+> = InnerTransaction<Arg, Pallet, Name, Asset, Ext>
 
-export type InnerTxEntry<
+export type Extensions<Ext> =
+  Ext extends Record<string, any>
+    ? Partial<Ext>
+    : Record<string, CustomSignedExtensionValues>
+
+export type TxEntry<
   Arg extends {} | undefined,
   Pallet extends string,
   Name extends string,
+  Ext,
   Asset,
 > = {
   /**
@@ -363,7 +387,7 @@ export type InnerTxEntry<
    */
   (
     ...args: Arg extends undefined ? [] : [data: Arg]
-  ): Transaction<Arg, Pallet, Name, Asset>
+  ): Transaction<Arg, Pallet, Name, Asset, Extensions<Ext>>
 }
 
 export type OfflineTxEntry<
@@ -397,13 +421,6 @@ export type OfflineTxEntry<
    */
   decodedCall: Enum<{ [P in Pallet]: Enum<{ [N in Name]: Arg }> }>
 }
-
-export type TxEntry<
-  Arg extends {} | undefined,
-  Pallet extends string,
-  Name extends string,
-  Asset,
-> = InnerTxEntry<Arg, Pallet, Name, Asset>
 
 export type TxFromBinary<Asset> = {
   /**