+++ packages/json-rpc/json-rpc-provider-proxy/src/internal-types.ts
@@ -4,8 +4,9 @@ import {
 } from "@polkadot-api/json-rpc-provider"
 import { InnerJsonRpcProvider } from "./public-types"
 
-export interface ConnectableJsonRpcConnection<T = any>
-  extends JsonRpcConnection {
+export interface ConnectableJsonRpcConnection<
+  T = any,
+> extends JsonRpcConnection {
   connect: (cb: InnerJsonRpcProvider<T>) => void
 }
 export type ReconnectableJsonRpcConnection = (