+++ packages/observable-client/src/chainHead/streams/index.ts
@@ -1,3 +1,4 @@
 export * from "./follow"
 export * from "./get-runtime-creator"
 export * from "./pinned-blocks"
+export * from "./pinned-blocks-types"