+++ packages/observable-client/src/chainHead/streams/pinned-blocks.ts
@@ -14,18 +14,11 @@ import { withStopRecovery } from "../enhancers"
 import type { FollowEvent } from "./follow"
 import { Runtime, getRuntimeCreator } from "./get-runtime-creator"
 import { BlockInfo } from "../chainHead"
-
-export interface PinnedBlock {
-  hash: string
-  number: number
-  parent: string
-  children: Set<string>
-  runtime: string
-  unpinnable: boolean
-  refCount: number
-  recovering: boolean
-  hasNewRuntime: boolean
-}
+import {
+  PinnedBlock,
+  PinnedBlocks,
+  PinnedBlockState,
+} from "./pinned-blocks-types"
 
 export interface BlockUsageEvent {
   type: "blockUsage"
@@ -33,38 +26,6 @@ export interface BlockUsageEvent {
 }
 interface CleanupEvent {
   type: "cleanup"
-}
-
-export const enum PinnedBlockState {
-  Initializing, // We are waiting to get ready on the "first" best-block event of the current or next chainHead follow subscription.
-  Ready, // Downstream events can be derived only when the PinnedBlockState is in this state
-  RecoveringInit, // There has been a stop-event and we are hopping that the first initialized event has an overlap with the previous blocks.
-  RecoveringFin, // We were not able to fully recover from the stop-even after the initialize event (the most recent block on init was behind the previous finalized one). So, we are waiting for a finalized event to be able to recover
-}
-
-type RecoveringInitState = { type: PinnedBlockState.RecoveringInit }
-type ReadyState = { type: PinnedBlockState.Ready }
-type InitializingState = {
-  type: PinnedBlockState.Initializing
-  pendingBlocks: Array<PinnedBlock>
-}
-type RecoveringFinState = {
-  type: PinnedBlockState.RecoveringFin
-  target: number
-  pendingBlocks: Array<PinnedBlock>
-}
-
-export type PinnedBlocks = {
-  best: string
-  finalized: string
-  runtimes: Record<string, Runtime>
-  blocks: Map<string, PinnedBlock>
-  finalizedRuntime: Runtime
-  state:
-    | RecoveringInitState
-    | RecoveringFinState
-    | ReadyState
-    | InitializingState
 }
 
 export const toBlockInfo = ({
@@ -156,6 +117,17 @@ export const getPinnedBlocks$ = (
         onUnpin(toUnpin)
       }
 
+      const tryToUnpin = (...target: PinnedBlock[]) => {
+        const finHeight = acc.blocks.get(acc.finalized)?.number ?? -1
+        const toRemove = target
+          .filter(
+            ({ refCount, recovering, number, pruned }) =>
+              !refCount && !recovering && (pruned || number < finHeight),
+          )
+          .map(({ hash }) => hash)
+        toRemove.length && unpinAndDelete(toRemove)
+      }
+
       switch (event.type) {
         case "initialized":
           if (
@@ -196,12 +168,10 @@ export const getPinnedBlocks$ = (
           let latestRuntime = acc.finalizedRuntime.codeHash
 
           event.finalizedBlockHashes.forEach((hash, i) => {
-            const unpinnable = i !== lastIdx
             const preexistingBlock = acc.blocks.get(hash)
 
             if (preexistingBlock) {
               preexistingBlock.recovering = false
-              preexistingBlock.unpinnable = unpinnable
             } else {
               const number = event.number + i
               const isNew = number > latestFinalizedHeight
@@ -222,7 +192,7 @@ export const getPinnedBlocks$ = (
                 children: new Set(
                   i === lastIdx ? [] : [event.finalizedBlockHashes[i + 1]],
                 ),
-                unpinnable,
+                pruned: false,
                 runtime: latestRuntime,
                 refCount: 0,
                 number,
@@ -263,10 +233,10 @@ export const getPinnedBlocks$ = (
             const block = {
               hash,
               number,
-              parent: parent,
+              parent,
               children: new Set<string>(),
-              runtime: event.newRuntime ? hash : parentNode.runtime,
-              unpinnable: false,
+              runtime: parentNode.runtime,
+              pruned: false,
               refCount: 0,
               recovering: false,
               hasNewRuntime: !!event.newRuntime,
@@ -274,6 +244,7 @@ export const getPinnedBlocks$ = (
             acc.blocks.set(hash, block)
             if (event.newRuntime) {
               const { codeHash } = event
+              block.runtime = codeHash!
               acc.runtimes[codeHash!] = getRuntime(codeHash!, hash)
             }
 
@@ -314,40 +285,23 @@ export const getPinnedBlocks$ = (
 
           event.prunedBlockHashes.forEach((hash) => {
             const block = acc.blocks.get(hash)
-            if (block) {
-              block.unpinnable = true
-            }
+            if (block) block.pruned = true
           })
 
-          let current = blocks.get(blocks.get(acc.finalized)!.parent)
-          while (current && !current.unpinnable) {
-            current.unpinnable = true
-            current = blocks.get(current.parent)
-          }
-
           if (shouldBeSetToReady) setToReady(acc)
-
           if (acc.state.type === PinnedBlockState.Ready) cleanup$.next()
 
           return acc
         }
         case "cleanup": {
-          const toUnpin = [...acc.blocks.values()]
-            .filter(({ unpinnable, refCount }) => unpinnable && !refCount)
-            .map(({ hash }) => hash)
-
-          unpinAndDelete(toUnpin)
+          tryToUnpin(...acc.blocks.values())
           return acc
         }
         case "blockUsage": {
-          if (!acc.blocks.has(event.value.hash)) return acc
-
-          const block = acc.blocks.get(event.value.hash)!
+          const block = acc.blocks.get(event.value.hash)
+          if (!block) return acc
           block.refCount += event.value.type === "hold" ? 1 : -1
-          if (block.refCount === 0 && !block.recovering && block.unpinnable) {
-            const toUnpin = [block.hash]
-            unpinAndDelete(toUnpin)
-          }
+          tryToUnpin(block)
           return acc
         }
       }