+++ packages/observable-client/src/utils/create-metadata-ctx.ts
@@ -65,6 +65,7 @@ export const createRuntimeCtx = (
     },
     accountId: AccountId(dynamicBuilder.ss58Prefix),
     getMortalityFromTx,
+    extVersions: metadata.extrinsic.version,
   }
 }
 
@@ -92,7 +93,7 @@ type DecodedExtrinsic = {
 )
 
 const mortalDecoder = enhanceDecoder(u16[1], (input) => {
-  const period = 2 << input % (1 << 4)
+  const period = 2 << (input % (1 << 4))
   const factor = Math.max(period >> 12, 1)
   const phase = (input >> 4) * factor
   return { mortal: true as const, period, phase }