+++ packages/ink-contracts/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+### Fixed
+
+- [performance] Cache `getInkClient` result.
+
+## 0.4.4 - 2025-12-08
+
+### Fixed
+
+- Revive event filtering when substrate-client is nested in other packages.
+
 ## 0.4.3 - 2025-11-08
 
 ### Fixed