+++ packages/known-chains/CHANGELOG.md
@@ -2,6 +2,25 @@
 
 ## Unreleased
 
+## 0.9.16 - 2025-12-08
+
+### Fixed
+
+- Update `lightSyncState`
+
+## 0.9.15 - 2025-11-24
+
+### Fixed
+
+- Update `lightSyncState`
+
+## 0.9.14 - 2025-11-19
+
+### Fixed
+
+- Update `lightSyncState`
+- Remove faulty bootnodes
+
 ## 0.9.13 - 2025-11-08
 
 ### Fixed