+++ packages/smoldot/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 0.3.15 - 2025-12-08
+
+### Fixed
+
+- Update smoldot to `2.0.40`
+
 ## 0.3.14 - 2025-09-18
 
 ### Fixed